Multiple changes to architecture of the extension and the way messages are sent from the content script (injected per page) and the background script (one instance per chrome).
Fixes multiple race conditions for changing the titles.

Adds a new badge state (temporary) in which badges stay on as long
as you keep the ctrl key pressed. Badges disappear when you release
the ctrl key. Note that if you press the ctrl key and release very
quickly the keyup event will not take effect because of the race
condition this patch fixes.

Maybe we can fix this by that if the ctrl key is released we don't
worry about the race condition. Shouldn't be a problem overall as
it just disables changing the titles for 350 ms (see
capture_keyboard_events.js).